Opened up the definition of Encoding a bit to avoid the need for useless getters/setters.
